Remove/Rename all occurences of iceberg.v2
