--- airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/integrations/destination/iceberg/v2/IcebergDestination.kt
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.iceberg.v2
-
-import io.airbyte.cdk.AirbyteDestinationRunner
-
-object IcebergDestination {
-    @JvmStatic
-    fun main(args: Array<String>) {
-        AirbyteDestinationRunner.run(*args)
-    }
-}
