--- airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/integrations/destination/iceberg/v2/io/Operation.kt
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.iceberg.v2.io
-
-/** Delta operations for data. */
-enum class Operation {
-    INSERT,
-    UPDATE,
-    DELETE,
-}
