+++ airbyte-integrations/connectors/destination-s3-data-lake/src/main/kotlin/io/airbyte/integrations/destination/s3_data_lake/IcebergDestination.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.integrations.destination.s3_data_lake
+
+import io.airbyte.cdk.AirbyteDestinationRunner
+
+object IcebergDestination {
+    @JvmStatic
+    fun main(args: Array<String>) {
+        AirbyteDestinationRunner.run(*args)
+    }
+}
