--- airbyte-integrations/connectors/destination-s3-data-lake/src/test-integration/kotlin/io/airbyte/integrations/destination/iceberg/v2/IcebergV2SpecTest.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.iceberg.v2
-
-import io.airbyte.cdk.load.spec.SpecTest
-
-class IcebergV2SpecTest : SpecTest()
